+++ actix-server/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## Unreleased
+
+### Changed
+
+* workers(0) defaults to num_cpus::get()
+
 ## [1.0.3] - 2020-05-19
 
 ### Changed